feat: refactor credit-related services to use utility functions instead of the credit service and add credit types for better organization
