+++ adaptive-app/src/types/index.ts
@@ -11,6 +11,7 @@ export type RouterInputs = inferRouterInputs<AppRouter>;
 export * from "./api_keys";
 // Re-export all domain-specific types for easy access
 export * from "./conversations";
+export * from "./credits";
 export * from "./messages";
 export * from "./organizations";
 export * from "./projects";
